files structure updated
