+++ controllers/users.js
@@ -1,0 +1,6 @@
+const usersRoute = (knexInstance) => (req, res,) => {
+    knexInstance.select('*').from('users')
+        .then(users => res.json(users))
+}
+
+export default usersRoute
